please help

+++ src/PointComparator.java
@@ -2,6 +2,6 @@ import java.util.Comparator;
 
 public class PointComparator implements Comparator<Node> {
     public int compare(Node n1, Node n2) {
-        return Double.compare(n1.getF(), n2.getF());
+        return Integer.compare(n1.getF(), n2.getF());
     }
 }
